init: repository with build pipeline
This uses an older revision of nixpkgs that contained glibc 2.34
This was a concious decision as compiling with too new of a glibc
provided a shared object that was unsuitable for distros which do
not update frequently (not bleeding edge).

In order to run this build (though you can use the releases instead),
run `nix develop`. Despite the cli call, it will build and copy over
the fixed lib to your current working directory.